add DeclareConfigMixin
